refacto(useGameState): extraire fonctions pures dans gameStateUtils (SRP)
- Créer gameStateUtils.ts : computeMultiplier, nextMilestone, generatePlayerName + constantes associées (COMBO_DECAY_MS, MULTIPLIER_THRESHOLDS, MILESTONES, MILESTONE_REPEAT_EVERY)
- useGameState.ts : remplacer les définitions locales par imports depuis gameStateUtils
- Ces fonctions pures ne dépendent d'aucun state/ref/hook React → testables sans composant monté
