+++ apps/playfield/src/hooks/useGameState.ts
@@ -13,6 +13,12 @@ import {
   onMusicGameOver,
 } from "../audio/pinballAudio";
 import { playfieldToScreenPercent, jitterScreenPoint } from "../utils/playfieldScreen";
+import {
+  computeMultiplier,
+  nextMilestone,
+  generatePlayerName,
+  COMBO_DECAY_MS,
+} from "./gameStateUtils";
 
 export type GameState = "idle" | "playing" | "game_over";
 
@@ -52,41 +58,6 @@ export interface ScoringCallbacks {
   onFeverEnd?: () => void;
 }
 
-const COMBO_DECAY_MS = 2000;
-const MULTIPLIER_THRESHOLDS = [5, 10, 20, 40] as const;
-
-const MILESTONES = [5_000, 15_000, 30_000];
-const MILESTONE_REPEAT_EVERY = 25_000; // au-delà de 50k
-
-// Plus haut seuil de {5k,15k,30k,50k,75k,100k,…} franchi entre prev et next.
-// Marque TOUS les seuils franchis dans `passed` (sinon les seuils intermédiaires
-// non retournés re-déclencheraient au prochain event) et renvoie le plus haut.
-function nextMilestone(prev: number, next: number, passed: Set<number>): number | null {
-  let crossed: number | null = null;
-  const mark = (m: number) => {
-    if (m > prev && m <= next && !passed.has(m)) {
-      passed.add(m);
-      if (crossed === null || m > crossed) crossed = m;
-    }
-  };
-  for (const m of MILESTONES) mark(m);
-  // Répétition tous les 25k au-delà de 50k.
-  for (let m = 50_000; m <= next; m += MILESTONE_REPEAT_EVERY) mark(m);
-  return crossed;
-}
-
-function computeMultiplier(combo: number): number {
-  if (combo < MULTIPLIER_THRESHOLDS[0]) return 1;
-  if (combo < MULTIPLIER_THRESHOLDS[1]) return 2;
-  if (combo < MULTIPLIER_THRESHOLDS[2]) return 3;
-  if (combo < MULTIPLIER_THRESHOLDS[3]) return 4;
-  return 5;
-}
-
-function generatePlayerName(): string {
-  const n = Math.floor(Math.random() * 10000).toString().padStart(4, "0");
-  return `PLAYER${n}`;
-}
 
 const initialBossHudEntry = (): BossHudEntry => ({
   active: false,
